Rename : 파일명 오타 수정
- assets 파일명 컨벤션 케밥케이스로 수정
